+++ src/initial_conditions/InitialConditions.jl
@@ -9,6 +9,7 @@ import ..Microphysics0Moment
 import ..Microphysics1Moment
 import ..Microphysics2Moment
 import ..Microphysics2MomentP3
+import ..QuadratureMicrophysics
 import ..PrescribedSST
 import ..SlabOceanSST
 import ..ᶜinterp